# docs/swagger.yaml
swagger: '2.0'
info:
  version: "1.0.0"
  title: STAFF WARS API
host: staffwars.azurewebsites.net
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json

tags:
  - name: "subordinate"
    description: 部下向けAPI
  - name: "boss"
    description: 上司向けAPI

paths:
  /api/subordinate/:
    post:
      tags:
        - subordinate
      summary: 部下の新規登録
      parameters:
        - name: body
          in : body
          required: true
          schema:
            type: object
            properties:
              name:
                title: 名前
                type: string
      responses:
        200:
          description: 成功
          schema:
            type: object
            properties:
              data:
                $ref: '#/definitions/Subordinate'
        default:
          description: エラー
          schema:
            $ref: '#/definitions/Error'
    get:
      tags:
        - subordinate
      summary: 部下の情報一覧
      responses:
        200:
          description: 成功
          schema:
            type: object
            properties:
              data:
                title: データ
                type: array
                items:
                  $ref: '#/definitions/Subordinate'
        default:
          description: エラー
          schema:
            $ref: '#/definitions/Error'
            
  /api/subordinate/{subordinate_id}/:
    post:
      tags:
        - subordinate
      summary: 部下の情報の更新
      parameters:
        - name: subordinate_id
          in: path
          required: true
          type: string
          default: 2
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              name:
                title: 名前
                type: string
                default: iwata-na
      responses:
        200:
          description: 成功
          schema:
            type: object
            properties:
              data:
                $ref: '#/definitions/Subordinate'
        default:
          description: エラー
          schema:
            $ref: '#/definitions/Error'
    get:
      tags:
        - subordinate
      summary: 部下の情報取得
      parameters:
        - name: subordinate_id
          in: path
          required: true
          type: string
          default: 2
      responses:
        200:
          description: 成功
          schema:
            type: object
            properties:
              data:
                $ref: '#/definitions/Subordinate'
        default:
          description: エラー
          schema:
            $ref: '#/definitions/Error'

  /api/boss/:
    post:
      tags:
        - boss
      summary: 上司の新規登録
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              id:
                title: 上司IDの
                type: string
                default: matsumura-s
              name:
                title: 名前
                type: string
                default: ショコ・マ・ツム−
      responses:
        200:
          description: 成功
          schema:
            type: object
            properties:
              data:
                $ref: '#/definitions/Boss'
        default:
          description: エラー
          schema:
            $ref: "#/definitions/Error"
    get:
      tags:
        - boss
      summary: 上司の情報一覧
      responses:
        200:
          description: 成功
          schema:
            type: object
            properties:
              data:
                title: データ
                type: array
                items:
                  $ref: '#/definitions/Boss'
        default:
          description: エラー
          schema:
            $ref: '#/definitions/Error'

  /api/boss/{boss_id}/:
    post:
      tags:
        - boss
      summary: 上司の情報の更新
      parameters:
        - name: boss_id
          in: path
          required: true
          type: string
          default: 1
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              name:
                title: 名前
                type: string
      responses:
        200:
          description: 成功
          schema:
            type: object
            properties:
              data:
                $ref: '#/definitions/Boss'
        default:
          description: エラー
          schema:
            $ref: '#/definitions/Error'
    get:
      tags:
        - boss
      summary: 上司の情報取得
      parameters:
        - name: boss_id
          in: path
          required: true
          type: string
          default: 1
      responses:
        200:
          description: 成功
          schema:
            type: object
            properties:
              data:
                $ref: '#/definitions/Boss'
        default:
          description: エラー
          schema:
            $ref: '#/definitions/Error'

  /api/boss/{boss_id}/regist/:
    post:
      tags:
        - boss
        - subordinate
      summary: 上司待ちを登録する
      parameters:
        - name: boss_id
          in: path
          required: true
          type: string
          default: 1
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              code:
                description: 部下コード(メールアドレスの@より前)
                title: 部下コード(メールアドレスの@より前)
                type: string
                default: iwata-na
      responses:
        200:
          description: 成功
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  $ref: "#/definitions/Subordinate"
        default:
          description: エラー
          schema:
            $ref: '#/definitions/Error'
    get:
      tags:
        - boss
      summary: 現在の上司待ち登録情報を取得する
      parameters:
        - name: boss_id
          in: path
          required: true
          type: string
          default: 1
      responses:
        200:
          description: 成功
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  $ref: "#/definitions/Subordinate"
        default:
          description: エラー
          schema:
            $ref: '#/definitions/Error'

  /api/boss/{boss_id}/unregist/:
    post:
      tags:
        - boss
        - subordinate
      summary: 上司待ちを解除する
      parameters:
        - name: boss_id
          in: path
          required: true
          type: string
          default: 1
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              code:
                description: 部下コード(メールアドレスの@より前)
                title: 部下コード(メールアドレスの@より前)
                type: string
                default: iwata-na

      responses:
        200:
          description: 成功
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  $ref: "#/definitions/Subordinate"
        default:
          description: エラー
          schema:
            $ref: '#/definitions/Error'

  /api/boss/{boss_id}/start/:
    post:
      tags:
        - boss
      summary: 早押し開始
      description: |
        リクエストを処理すると、MilkcocoaでPush通知がくる。通知のTypeに応じて処理が分かれる
        notice => カウントダウン開始時刻の通知
        pre_start => カウントダウン開始５秒前の通知
        start => カウントダウン開始の通知
        pre_finish => カウントダウン終了5秒前の通知
        finish => カウントダウン終了の通知
        result => 早押し結果の通知
      parameters:
        - name: boss_id
          in: path
          required: true
          type: string
          default: 1
      responses:
        200:
          description: 成功
          schema:
            type: object
            properties:
              data:
                title: 開始時刻
                type: string
        default:
          description: エラー
          schema:
            $ref: '#/definitions/Error'

  /api/boss/{boss_id}/push/:
    post:
      tags:
        - subordinate
      summary: 部下が早押しボタンを押した
      consumes:
        - application/json
      parameters:
        - name: boss_id
          in: path
          required: true
          type: string
          default: 1
        -
          name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              code:
                description: 部下コード(メールアドレスの@より前)
                title: 部下コード(メールアドレスの@より前)
                type: string
                default: iwata-na
      responses:
        200:
          description: 成功
        default:
          description: エラー
          schema:
            $ref: '#/definitions/Error'

definitions:
  Error:
    title: エラー情報
    type: object
    properties:
      message:
        type: string
        title: エラーメッセージ
      error:
        type: object
        title: エラー内容

  Subordinate:
    title: 部下情報
    type: object
    properties:
      id:
        type: string
        title: 部下ID
        default: 2
      code:
        type: string
        title: コード(メールアドレスの@より前の部分)
        default: iwata-na
      name:
        type: string
        title: 名前
        default: イワンコフ
      organization:
        type: string
        title: 所属
        default: 全社員早押上司争奪戦推進機関
      icon:
        type: string
        title: アイコン画像のURL
        default: http://staffwars.azurewebsites.net/icon/

  Boss:
    title: 上司情報
    type: object
    properties:
      id:
        type: string
        title: 上司ID
      code:
        type: string
        title: コード(メールアドレスの@より前の部分)
      name:
        type: string
        title: 名前
      organization:
        type: string
        title: 所属
      start_datetime:
        type: string
        title: 開始予定時刻
      registry:
        title: 待っている部下達
        type: array
        items:
          $ref: '#/definitions/Subordinate'
      push_id:
        title: 早押し大会を区別するためのID
        description: 早押し大会を実行する度にインクリメントする
        type: number
      push_list:
        title: 最後の早押し結果
        type: array
        items:
          type: object
          title: ランキング順データ
          properties:
            datetime:
              type: string
              title: 押した日時
            subordinate:
              $ref: '#/definitions/Subordinate'